refactor(mock): convert to TS
